move log_analyzer to utils
